pycon_graphql/requirements.txt
aniso8601==3.0.2
atomicwrites==1.2.1
attrs==18.2.0
coverage==4.5.2
Django==2.1.5
django-simplemde==0.1.3
django-graphql-extensions==0.0.6
django-graphql-jwt==0.2.0
factory-boy==2.11.1
Faker==1.0.2
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
more-itertools==5.0.0
Pillow==5.4.1
pluggy==0.8.1
promise==2.2.1
py==1.7.0
pytest==4.1.1
pytest-cov==2.6.1
pytest-django==3.4.5
python-dateutil==2.7.5
pytz==2018.9
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
text-unidecode==1.2
